pyproject.toml
[project]
name = "mctrack"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "easydict>=1.13",
    "lap>=0.5.12",
    "motmetrics==1.1.3",
    "numba>=0.58.1",
    "nuscenes-devkit>=1.1.11",
    "pandas==1.5.0",
    "pillow>=10.4.0",
    "pyyaml>=6.0.2",
]
